tests(*) make proxy client usage more robust
### Summary

It seems like the proxy client when used multiple times
for subsequent request may be a bit flaky at times.
Especially if the previous request errored, it may
have closed the connectiont too, thus subsequent
request failing on that same client.

This commit grabs a new client for each request.

This should be bullet-proofed on Kong repo, but this
fixes the issues here for now.